requirements.txt
asgiref==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
dj-database-url==0.5.0
Django==2.2.8
django-allauth==0.41.0
django-cors-headers==3.7.0
django-email-verification==0.1.0
django-environ==0.4.5
django-filter==1.1.0
django-heroku==0.3.1
django-phone-verify==2.0.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-rest-authemail==2.0.6
django-rest-registration==0.6.3
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
django-elasticsearch-dsl
djangoflutterwave
gunicorn==20.1.0
idna==2.10
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
nexmo==2.5.2
oauthlib==3.1.0
openapi-codec==1.3.2
phonenumbers==8.12.24
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.1.2
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.1
twilio==6.50.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
